refactor: :recycle: export apollo client conn
